+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.3.0
 
 - Update to SlmQueue 0.3
+- [BC] BeanstalkdQueueInterface have been renamed to TubeInterface, BeanstalkdQueue have been renamed to Tube
 
 # 0.2.1